requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Task queue
celery[redis]==5.3.6
redis==5.0.4

# Settings management
pydantic==2.7.1
pydantic-settings==2.3.0

# STAC / raster
pystac-client==0.7.7
planetary-computer==1.0.0
stackstac==0.5.1
rasterio==1.3.10
rioxarray==0.15.5
xarray==2024.3.0
numpy==1.26.4

# Dask (chunked computation)
dask[array]==2024.4.2
distributed==2024.4.2

# Geospatial
geopandas==0.14.4
shapely==2.0.4
pyproj==3.6.1

# Image preview
Pillow==10.3.0

# Utilities
httpx==0.27.0
python-dateutil==2.9.0
orjson==3.10.3
aiofiles==23.2.1
